🔨 refactor medium chart component
