+++ site/owid.scss
@@ -53,9 +53,10 @@
 @import "./search/SearchTopicsRefinementList.scss";
 @import "./search/_DEPRECATEDChartHit.scss";
 @import "./search/SearchHorizontalDivider.scss";
+@import "./search/SearchChartHitCaptionedLink.scss";
 @import "./search/SearchChartHitHeader.scss";
 @import "./search/SearchChartHitThumbnail.scss";
-@import "./search/SearchChartHitMedium.scss";
+@import "./search/SearchChartHitRichData.scss";
 @import "./search/SearchChartHitSmall.scss";
 @import "./search/SearchChartHitDataDisplay.scss";
 @import "./search/SearchChartHitDataTable.scss";
