+++ site/search/SearchChartHitCaptionedLink.scss
@@ -1,0 +1,29 @@
+.search-chart-hit-captioned-link {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: flex-end;
+    gap: 8px;
+    height: 100%;
+
+    @include md-down {
+        scroll-snap-align: start;
+    }
+
+    &:hover {
+        .search-chart-hit-captioned-link__label {
+            text-decoration: none;
+        }
+    }
+
+    .search-chart-hit-captioned-link__label {
+        @include note-1-regular;
+        color: $blue-90;
+        text-decoration: underline;
+        text-underline-offset: 3px;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+        max-width: 100%;
+    }
+}
