+++ site/search/SearchChartHitCaptionedLink.tsx
@@ -1,0 +1,28 @@
+import cx from "classnames"
+
+export function CaptionedLink({
+    url,
+    className,
+    onClick,
+    children,
+    caption,
+}: {
+    url: string
+    className?: string
+    onClick?: () => void
+    children: React.ReactNode
+    caption: string
+}): React.ReactElement {
+    return (
+        <a
+            href={url}
+            className={cx("search-chart-hit-captioned-link", className)}
+            onClick={onClick}
+        >
+            {children}
+            <div className="search-chart-hit-captioned-link__label">
+                {caption}
+            </div>
+        </a>
+    )
+}
